requirements.txt
scipy
einops
pyyaml
huggingface_hub
encodec
matplotlib
transformers
pytorch-lightning
tensorboardX
soundfile
numpy
jsonargparse
torchcrepe
librosa
pesq
inflect
loguru
polars
natsort
tqdm
requests
sounddevice